Ecobee Usability Updates
Redid the functions in the Ecobee_1-3 script to make a single function
for setting the thermostat, and a single setting for reading values from
the thermostat.

More usability updates are to come, though they will break any work done
using the two functions described above.  On the todo list is:
Make things more robust in several spots, sanitize inputs better
Explicitly check if the APP needs to be authorized
Make the authorization process a bit more clear
More comments and docstrings
Clean up the mess of strings that are everywhere